Porting everything to use Double arrays to avoid costly float to double conversion on the jvm

+++ audio_be/src/commonMain/kotlin/KlangAudioRenderer.kt
@@ -67,7 +67,7 @@ class KlangAudioRenderer(
         val right = mix.right
 
         // Cache max value to avoid repeated field access/casting
-        val maxShort = Short.MAX_VALUE.toFloat()
+        val maxShort = Short.MAX_VALUE
 
         for (i in 0 until blockFrames) {
             val lSample = left[i]
@@ -77,17 +77,17 @@ class KlangAudioRenderer(
             // Most samples are within safe bounds [-1.0, 1.0].
             // We skip all math for them to preserve CPU and Transparency (Unity Gain).
 
-            val lOut = if (lSample >= -1.0f && lSample <= 1.0f) {
+            val lOut = if (lSample >= -1.0 && lSample <= 1.0) {
                 (lSample * maxShort).toInt()
-            } else if (lSample > 1.0f) {
+            } else if (lSample > 1.0) {
                 Short.MAX_VALUE.toInt()
             } else {
                 Short.MIN_VALUE.toInt()
             }
 
-            val rOut = if (rSample >= -1.0f && rSample <= 1.0f) {
+            val rOut = if (rSample >= -1.0 && rSample <= 1.0) {
                 (rSample * maxShort).toInt()
-            } else if (rSample > 1.0f) {
+            } else if (rSample > 1.0) {
                 Short.MAX_VALUE.toInt()
             } else {
                 Short.MIN_VALUE.toInt()
